+++ src/inmanta/protocol/common.py
@@ -440,6 +440,7 @@ class MethodProperties:
         self.function = function
         self._varkw: bool = varkw
         self._varkw_name: Optional[str] = None
+        self._return_type: Optional[type] = None
 
         self._parsed_docstring = docstring_parser.parse(text=function.__doc__, style=docstring_parser.DocstringStyle.REST)
         self._docstring_parameter_map = {p.arg_name: p.description for p in self._parsed_docstring.params}
@@ -460,6 +461,12 @@ class MethodProperties:
     @property
     def enforce_auth(self) -> bool:
         return self._enforce_auth
+
+    @property
+    def return_type(self) -> type:
+        if self._return_type is None:
+            raise InvalidMethodDefinition("Only typed methods have a return type")
+        return self._return_type
 
     def validate_arguments(self, values: dict[str, Any]) -> dict[str, Any]:
         """
@@ -551,6 +558,7 @@ class MethodProperties:
                 f"A key/value argument is only allowed when `varkw` is set to true in method {self.function}."
             )
 
+        self._return_type = type_hints["return"]
         self._validate_return_type(type_hints["return"], strict=self.strict_typing)
 
     def _validate_return_type(self, arg_type: type, *, strict: bool = True) -> None:
