+++ tests/protocol/test_typed_client.py
@@ -23,7 +23,7 @@ from utils import configure
 
 
 async def test_local_client(unused_tcp_port: int, postgres_db: PostgresProc, database_name: str, async_finalizer) -> None:
-    """Test methods with simple return types"""
+    """The the local clien"""
     configure(unused_tcp_port, database_name, postgres_db.port)
 
     class ProjectServer(ServerSlice):
